feat(server): fields should be good going for relations
