+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/workspace-migration-v2-field-actions-builder.ts
@@ -14,20 +14,11 @@ import {
   getWorkspaceMigrationV2FieldCreateAction,
   getWorkspaceMigrationV2FieldDeleteAction,
 } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/get-workspace-migration-v2-field-actions';
-import {
-  UniqueIdentifierWorkspaceMigrationObjectInputMapDispatcher,
-  UpdatedObjectMetadataFieldAndRelationMatrix,
-} from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/workspace-migration-builder-v2.service';
 import { transformMetadataForComparison } from 'src/engine/workspace-manager/workspace-sync-metadata/comparators/utils/transform-metadata-for-comparison.util';
 
-import { isRelationFieldMetadataType } from 'src/engine/utils/is-relation-field-metadata-type.util';
 import { FromTo } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-action-common-v2';
-import { WorkspaceMigrationObjectInput } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-object-input';
+import { UpdatedObjectMetadataFieldAndRelationMatrix } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/compute-updated-object-metadata-field-and-relations-delete-created-updated-matrix.util';
 import { compareFieldMetadataInputAndGetUpdateFieldActions } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/workspace-migration-field-metadata-input-comparator.util';
-import {
-  CustomDeletedCreatedUpdatedMatrix,
-  deletedCreatedUpdatedMatrixDispatcher,
-} from './utils/deleted-created-updated-matrix-dispatcher.util';
 
 const shouldNotOverrideDefaultValue = (type: FieldMetadataType) => {
   return [
@@ -85,35 +76,6 @@ export const compareTwoWorkspaceMigrationFieldInput = ({
   return fieldMetadataDifference;
 };
 
-type DeletedCreatedUpdatedFieldInputMatrix = {
-  objectMetadataInput: WorkspaceMigrationObjectInput; // Should omit fields and fieldsInputs
-} & CustomDeletedCreatedUpdatedMatrix<
-  'fieldMetadata',
-  WorkspaceMigrationFieldInput
->;
-
-const computeUpdatedObjectMetadataFieldMatriceDispatcher = (
-  updatedObjectMetadata: UniqueIdentifierWorkspaceMigrationObjectInputMapDispatcher['updated'],
-): DeletedCreatedUpdatedFieldInputMatrix[] => {
-  const matriceAccumulator: DeletedCreatedUpdatedFieldInputMatrix[] = [];
-
-  for (const { from, to } of updatedObjectMetadata) {
-    const matrixResult = deletedCreatedUpdatedMatrixDispatcher({
-      from: from.fieldInputs,
-      to: to.fieldInputs,
-    });
-
-    matriceAccumulator.push({
-      objectMetadataInput: to,
-      createdFieldMetadata: matrixResult.created,
-      deletedFieldMetadata: matrixResult.deleted,
-      updatedFieldMetadata: matrixResult.updated,
-    });
-  }
-
-  return matriceAccumulator;
-};
-
 export const buildWorkspaceMigrationV2FieldActions = (
   objectMetadataDeletedCreatedUpdatedFields: UpdatedObjectMetadataFieldAndRelationMatrix[],
 ): WorkspaceMigrationFieldActionV2[] => {
@@ -126,20 +88,21 @@ export const buildWorkspaceMigrationV2FieldActions = (
     updatedFieldMetadata,
     objectMetadataInput,
   } of objectMetadataDeletedCreatedUpdatedFields) {
-    // make more readable
-    const updateFieldActions = updatedFieldMetadata
-      .filter(({ to: toField }) => !isRelationFieldMetadataType(toField.type))
-      .flatMap(({ from, to }) =>
-        compareFieldMetadataInputAndGetUpdateFieldActions({
+    const updateFieldActions = updatedFieldMetadata.map<UpdateFieldAction>(
+      ({ from, to }) => {
+        const updates = compareFieldMetadataInputAndGetUpdateFieldActions({
           from,
           to,
+        });
+
+        return {
+          type: 'update_field',
+          fieldMetadataInput: to,
           objectMetadataInput,
-        }),
-      )
-      .map<UpdateFieldAction>((updateFieldAction) => ({
-        ...updateFieldAction,
-        type: 'update_field',
-      }));
+          updates,
+        };
+      },
+    );
 
     const createFieldAction = createdFieldMetadata.map((fieldMetadataInput) =>
       getWorkspaceMigrationV2FieldCreateAction({
