+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/workspace-migration-v2-relation-actions-builder.ts
@@ -1,16 +1,14 @@
-import { isRelationFieldMetadataType } from 'src/engine/utils/is-relation-field-metadata-type.util';
 import {
   UpdateRelationAction,
   WorkspaceMigrationRelationActionV2,
 } from 'src/engine/workspace-manager/workspace-migration-v2/types/workspace-migration-relation-action-v2';
+import { UpdatedObjectMetadataFieldAndRelationMatrix } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/compute-updated-object-metadata-field-and-relations-delete-created-updated-matrix.util';
 import {
   getWorkspaceMigrationV2RelationCreateAction,
   getWorkspaceMigrationV2RelationDeleteAction,
 } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/get-workspace-migration-v2-relation-actions';
-import { compareFieldMetadataInputAndGetUpdateFieldActions } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/utils/workspace-migration-field-metadata-input-comparator.util';
-import { UpdatedObjectMetadataFieldAndRelationMatrix } from 'src/engine/workspace-manager/workspace-migration-v2/workspace-migration-builder-v2/workspace-migration-builder-v2.service';
 
-export const buildWorkspaceMigrationV2FieldActions = (
+export const buildWorkspaceMigrationV2RelationActions = (
   objectMetadataDeletedCreatedUpdatedFields: UpdatedObjectMetadataFieldAndRelationMatrix[],
 ): WorkspaceMigrationRelationActionV2[] => {
   let allUpdatedObjectMetadataFieldActions: WorkspaceMigrationRelationActionV2[] =
@@ -23,19 +21,7 @@ export const buildWorkspaceMigrationV2FieldActions = (
     objectMetadataInput,
   } of objectMetadataDeletedCreatedUpdatedFields) {
     // make more readable
-    const updateFieldActions = updatedRelationFieldMetadata
-      .filter(({ to: toField }) => !isRelationFieldMetadataType(toField.type))
-      .flatMap(({ from, to }) =>
-        compareFieldMetadataInputAndGetUpdateFieldActions({
-          from,
-          to,
-          objectMetadataInput,
-        }),
-      )
-      .map<UpdateRelationAction>(() => ({
-        type: 'update_relation',
-        // TODO
-      }));
+    const updateFieldActions: UpdateRelationAction[] = [];
 
     const createFieldAction = createdRelationFieldMetadata.map(
       (fieldMetadataInput) =>
